Ambrose zen

+++ Repository/Impl/ResourceRepository.cs
@@ -8,14 +8,13 @@ namespace Repository.Impl;
 
 public class ResourceRepository : IResourceRepository
 {
-    
     private readonly BaseDao<Resource> _resourceDao;
-    
+
     public ResourceRepository(DocumentManagementSystemDbContext context)
     {
         _resourceDao = new BaseDao<Resource>(context ?? throw new ArgumentNullException(nameof(context)));
     }
-    
+
     public async Task AddRangeAsync(List<ResourceDto> resources)
     {
         if (resources == null || resources.Count == 0)
@@ -26,30 +25,27 @@ public class ResourceRepository : IResourceRepository
 
         // Tìm API mới chưa có trong DB
         var newResources = resources
-            .Where(api => !existingResources.Any(r => r.ResourceApi == api.ResourceApi /*&& r.ResourceMethod == api.ResourceMethod*/))
+            .Where(api =>
+                !existingResources.Any(
+                    r => r.ResourceApi == api.ResourceApi /*&& r.ResourceMethod == api.ResourceMethod*/))
             .Select(api => new Resource
             {
                 ResourceId = Guid.NewGuid(),
                 ResourceName = api.ResourceName,
                 ResourceApi = api.ResourceApi,
-                PermissionId = api.PermissionId,
+                PermissionId = api.PermissionId
             })
             .ToList();
 
-        if (newResources.Count > 0)
-        {
-            await _resourceDao.AddRangeAsync(newResources);
-        }
+        if (newResources.Count > 0) await _resourceDao.AddRangeAsync(newResources);
 
         // (Optional) Xóa API cũ nếu không còn tồn tại trong danh sách mới
         var removedResources = existingResources
-            .Where(r => !resources.Any(api => api.ResourceApi == r.ResourceApi /*&& api.ResourceMethod == r.ResourceMethod*/))
+            .Where(r => !resources.Any(
+                api => api.ResourceApi == r.ResourceApi /*&& api.ResourceMethod == r.ResourceMethod*/))
             .ToList();
 
-        if (removedResources.Count > 0)
-        {
-            await _resourceDao.RemoveRangeAsync(removedResources);
-        }
+        if (removedResources.Count > 0) await _resourceDao.RemoveRangeAsync(removedResources);
     }
 
     public async Task AddAsync(Resource resource)
@@ -68,11 +64,10 @@ public class ResourceRepository : IResourceRepository
         if (resourceApi == null) throw new ArgumentNullException(nameof(resourceApi));
         return await _resourceDao.FindByAsync(p => p.ResourceApi == resourceApi);
     }
-    
+
     public async Task<Resource?> FindResourceByIdAsync(Guid? resourceId)
     {
         if (resourceId == null) throw new ArgumentNullException(nameof(resourceId));
         return await _resourceDao.FindByAsync(p => p.ResourceId == resourceId);
     }
-    
 }
